+++ lib/domain/enums/enums_declarations.dart
@@ -1,15 +1,26 @@
-abstract interface class EnumLabel {
-  String get label;
-}
+// abstract interface class EnumLabel {
+//   String get label;
+// }
 
-enum Sex implements EnumLabel {
+enum Sex {
   male('Masculino'),
   female('Feminino');
   // other('Outro');
 
-  @override
   final String label;
   const Sex(this.label);
 
   static Sex byName(String name) => values.byName(name);
+}
+
+enum AttachmentType {
+  pdf('pdf'),
+  text('texto'),
+  image('imagem'),
+  other('outro');
+
+  final String label;
+  const AttachmentType(this.label);
+
+  static AttachmentType byName(String name) => values.byName(name);
 }
